A base class for CNode and MNode.

D1Node is a base class for CNode and MNode classes and contains class slots and methods that are common between these two child classes.

Slots

identifier

A character string containing a URN that uniquely identifiers the node

name

A character string containing a plain text name for the node

description

A character string describing the node

baseURL

A character string of the registered baseURL for the node, which does not include the version string

subject

A character string containing the Distinguished Name of this node, used for authentication

contactSubject

The Distinguished Name of contact person for this node

replicate

A logical flag indicating whether the node accepts replicas

type

The node type, either 'mn' or 'cn'

state

A character string that indicates whether or not the node is accessible, either 'up' or 'down'

services

A data.frame containing the service tiers supported by this node.

serviceUrls

A data.frame that contains DataONE service Urls

APIversion

A character string indicating version of the DataONE API for this node, e.g. "v2"

env

A character string, either 'prod' if this node is in the production environment, otherwise 'test'

Methods

D1Node-initialize{initialize}: Initialize a D1Node

D1Node: Create a MNode object representing a DataONE Member Node repository.

archive: Change the state of an object so that it is hidden from searches.

describeObject: Get header information for a given pid.

getChecksum: Get the checksum for the data object associated with the specified pid.

getObject: Get the bytes associated with an object on a node.

getQueryEngineDescription: Query a node for the list of query engines available on the node.

getSystemMetadata: Get the metadata describing system properties associated with an object on the Node.

listObjects: Retrieve the list of objects that match the search parameters.

listQueryEngines: Query a node for the list of query engines available on the node.

ping: Test if a node is online and accepting DataONE requests.

encodeSolr: Encode the input for Solr Queries.

query: Search DataONE for data and metadata objects.

isAuthorized: Check if an action is authorized for the specified identifier.