Deprecated

The following items are deprecated in this release of dataone and will be removed in a future release

These S4 methods are deprecated

setObsoletedBy: Set a pid as being obsoleted by another pid.

CertificateManager: Create a CertificateManager object.

getCertLocation: Get the file path on disk of the client certificate file.

showClientSubject: Get DataONE Identity as Stored in the CILogon Certificate.

isCertExpired: Determine if an X.509 certificate has expired.

getCertExpires: Show the date and time when an X.509 certificate expires.

downloadCert: Open the CILogon Certificate download page in the default browser.

obscureCert: Obscure the CILogon Client Certificate.

restoreCert: Restore the CILogon client certificate by renaming it to its original location