Evaluate DataONE authentication.

A valid DataONE authentication method is looked for and all authentication information is retrieved from it.

evaluateAuth(.Object, ...)

# S4 method for AuthenticationManager
evaluateAuth(.Object, node)

Arguments

.Object
an Authentication Object.

...
additional parameters

node
A D1Node object.

Value

A hash containing authentication information.

Details

If the node specified in the 'node' parameter is a DataONE v2 node or higher, then an authentication token is checked if one exists. If it is readable and not expired, then information for the token is returned. If a valid token does not exist, then the X.509 certificate is checked, if it exists. If it is valid then information is returned for the certificate.