Query a node for the list of query engines available on the node

Query a node for the list of query engines available on the node

getQueryEngineDescription(x, ...)

# S4 method for D1Node
getQueryEngineDescription(x, queryEngineName)

Arguments

x

The CNode or MNode to query

...

(Additional arguments - not yet used.)

queryEngineName

The query engine name to get a description for.

Value

list The query engine description

Examples

if (FALSE) {
library(dataone)
cn <- CNode("PROD")
engineDesc <- getQueryEngineDescription(cn, "solr")
cat(sprintf("Query engine version: %s\n", engineDesc$queryEngineVersion))
cat(sprintf("Query engine name: %s\n", engineDesc$name))
engineDesc <- getQueryEngineDescription(cn, "solr")
head(engineDesc$queryFields, n=3L)
}