Retrieve the list of objects that match the search parameters

Retrieve the list of objects that match the search parameters

listObjects(x, ...)

# S4 method for D1Node
listObjects(
  x,
  fromDate = as.character(NA),
  toDate = as.character(NA),
  formatId = as.character(NA),
  replicaStatus = as.logical(TRUE),
  start = as.integer(0),
  count = as.integer(1000)
)

Arguments

x

The Node instance from which the SystemMetadata will be downloaded

...

(Not yet used.)

fromDate

Entries with a modified date greater than 'fromDate' will be returned. This value must be specified in ISO 8601 format, i.e. "YYYY-MM-DDTHH:MM:SS.mmm+00:00"

toDate

Entries with a modified date less than 'toDate' will be returned. This value must be specified in ISO 8601 format, i.e. "YYYY-MM-DDTHH:MM:SS.mmm+00:00"

formatId

The format to match, for example "eml://ecoinformatics.org/eml-2.1.1"

replicaStatus

A logical value that determines if replica (object not on it's origin node) should be returned. Default is TRUE.

start

An integer that specifies the first element of the result set that will be returned

count

An integer that specifies how many results will be returned

Value

list Objects that met the search criteria

list Objects that met the search criteria

Details

The list of objects that is returned is paged according to the 'start' and 'count' values, so that large result sets can be returned over multiple calls.

See also

Examples

if (FALSE) {
library(dataone)
cn <- CNode("STAGING")
fromDate <- "2013-01-01T01:01:01.000+00:00"
toDate <- "2015-12-31T01:01:01.000+00:00"
formatId <- "eml://ecoinformatics.org/eml-2.1.0"
start <- 0
count <- 5
objects <- listObjects(cn, fromDate=fromDate, toDate=toDate, 
    formatId=formatId, start=start, count=count)
# Inspect id of first object 
objects[1]$objectInfo$identifier
}