Test if a node is online and accepting DataONE requests

Test if a node is online and accepting DataONE requests

ping(x, ...)

# S4 method for D1Node
ping(x)

Arguments

x
The CNode or MNode to check

...
(Not yet used)

Value

logical A logical value set to TRUE if the node is up and FALSE if it is not

logical A logical value set to TRUE if the node is up and FALSE if it is not

Examples

if (FALSE) {
cn <- CNode()
mn <- getMNode(cn, "urn:node:KNB")
isAlive <- ping(mn)
}